+++ Solution.h
@@ -20,8 +20,8 @@ public: //TROCAR PARA PRIVADO ?
 
 	bool CostCalculated;
 	int num_items_in_sol;
-	double cost;
-	double used_capacity;
+	int cost;
+	int used_capacity;
 	unsigned long hash;
 	ProblemInstance* p;
 	vector<bool> inside;
@@ -42,7 +42,7 @@ public:
 	~Solution();
 
 	int get_size();
-	double getCost();
+	int getCost();
 
 	void add_item(int i);
 	void add_itemO(int i);
